smurf: support huge arrays
- Move to using 64 bit NDF API
- Change dim_t type from unsigned to signed 64 bit int
- Use dim_t instead of hdsdim (dim_t is now an alias for hdsdim)
- Change most size_t values to dim_t
- Fixes for all compiler warnings

Some of the warnings (particularly warnings to do with casts that remove
const) are generated by code that needs to remain unchanged. In such
cases the gcc warnings are suppressed by inserting '#pragma GCC
diagnostic ignored "-Wxxx"' immediately before the lines of code
concerned, with a corresponding '#pragma GCC diagnostic pop' after the
code. Unused function parameters are marked using the gcc "unused"
attribute, rather than being removed.

+++ applications/smurf/libsmf/smf_jsatilei2xy.c
@@ -136,7 +136,7 @@ void smf_jsatilei2xy( int itile, smfJSATiling *skytiling, int *xt, int *yt,
    if( itile >= 0 && itile < 12*nsq ){
 
 /* Get the zero-based index of the facet contained the tile. */
-      fi = itile / nsq;
+      fi = (int) ( itile / nsq );
 
 /* Get the offset in tiles into this facet of the required tile. */
       tj = itile - fi*nsq;
